Feature/swap message with image

+++ src/events/message.js
@@ -6,9 +6,9 @@ const { getTags } = require('../boot');
 const retries = new Map();
 
 function handleRetry(author, retryThreshold = config('bot.msgRetryThreshold')) {
-  const retryCount = retries.get(author.username) + 1;
+  const retryCount = retries.get(author.id) + 1;
 
-  retries.set(author.username, retryCount);
+  retries.set(author.id, retryCount);
 
   if (retryCount >= retryThreshold * 2) {
     const attachment = new Attachment(getReservedImage('angryPepe'));
@@ -34,52 +34,73 @@ function throttleUser(
   callback,
   throttleTime = config('bot.msgThrottleTime'),
 ) {
-  if (retries.has(author.username)) {
+  if (retries.has(author.id)) {
     handleRetry(author);
-    return;
+    throw Error('Preventing flood.');
   }
 
   setInterval(() => {
-    retries.delete(author.username);
+    retries.delete(author.id);
   }, throttleTime * 1000);
 
-  retries.set(author.username, 1);
+  retries.set(author.id, 1);
 
   callback();
 }
 
 module.exports = async message => {
-  if (message.author.bot) {
+  const { author, client, content, channel } = message;
+
+  if (author.bot) {
     return;
   }
 
-  const botUsername = message.client.user.toString();
+  const botUsername = client.user.toString();
 
-  if (message.content.startsWith(botUsername)) {
-    const [username, tag, ...gibberish] = message.content.trim().split(' ');
-
-    if (gibberish.length > 0) {
-      const attachment = new Attachment(getReservedImage('angryPepe'));
-
-      message.channel.send('**EI _BORDA_, UMA TAG DE CADA VEZ, OK?!?!**', {
-        files: [attachment],
-      });
-
-      return;
-    }
-
-    if (!tag) {
-      message.author.send(`Exemplo :: _${username} {tag}_`);
-
-      return;
-    }
-
+  if (content.startsWith(botUsername)) {
     const allowedTags = [...getTags().keys()].filter(
       key => key !== '__reserved',
     );
 
-    if (!allowedTags.includes(tag)) {
-      message.author.send(`Tags :: ${allowedTags.join(' , ')}`);
+    const [, tag, ...gibberish] = content.trim().split(' ');
+
+    if (gibberish.length > 0) {
+      const attachment = new Attachment(getReservedImage('angryPepe'));
+
+      try {
+        const warned = await author.send(
+          '**EI _BORDA_, UMA TAG DE CADA VEZ, OK?!?!**',
+          { files: [attachment] },
+        );
+
+        if (warned) {
+          await author.send(`Tags :: ${allowedTags.join(' , ')}`);
+        }
+
+        if (message.deletable) {
+          await message.delete();
+        }
+      } catch (e) {
+        console.log('WARNING :: ', e.message);
+      }
+
+      return;
+    }
+
+    if (!tag || !allowedTags.includes(tag)) {
+      try {
+        if (message.deletable) {
+          const deleted = await message.delete();
+
+          if (deleted) {
+            throttleUser(author, async () => {
+              author.send(`Tags :: ${allowedTags.join(' , ')}`);
+            });
+          }
+        }
+      } catch (e) {
+        console.log('WARNING :: ', e.message);
+      }
 
       return;
     }
@@ -87,14 +108,27 @@ module.exports = async message => {
     try {
       const image = getImage(tag);
 
-      throttleUser(message.author, async () => {
-        const deleted = await message.delete();
-        if (deleted) {
-          await message.channel.send(new Attachment(image));
+      throttleUser(author, async () => {
+        const attachment = new Attachment(image);
+
+        try {
+          if (message.editable) {
+            await message.edit(attachment);
+          } else if (message.deletable) {
+            await message.delete();
+          }
+
+          await channel.send(attachment);
+        } catch (e) {
+          console.log('WARNING :: ', e.message);
         }
       });
     } catch (e) {
-      console.log(e);
+      console.log('WARNING :: ', e.message);
+
+      if (message.deletable) {
+        await message.delete();
+      }
     }
   }
 };
